refactor: extract security utilities from server.py
